ai-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
faiss-cpu==1.7.4
sentence-transformers==2.7.0
numpy==1.26.4
pydantic==2.6.4